Found a way to use init_array in wasm after all.
Except its really fiddly to get working because there must be a single
unique section with the form:

 1. .section .init_array[.N]
 2. .p2align 2, 0x0
 3. <sequence of .int32 data directives>

When lowered from llvm ir, multiple sections with 1 and 3 are helpfully
merged into a 1 and several 3, but as soon as one adds alignment it
breaks merger with any previous `.section`s

As such I put in place a required early call to wasm_init to do 1 and 2
so that all the subsequent calls can do 1 and 3 which gets beautifully
merged into 1 2 3*.

At least until we stop emitting llvm-ir or asm at which point it breaks.
Since neither ir nor asm are being generated at this point its hard to
pinpoint the exact reason, but the hypothesis is that this is working
until the number of codegen units processing the code is more than one
and then probably some of the sections become incomplete or otherwise
deficient in their respective processors.

One potential mitigation for this is to make use of the priority
function which lets one specify a u16 integer to give an initialization
order to the constructors. Since these are independent sections they
should be able to be emitted independently with no problem, but then
this re-opens the previous unsolved problem of how to get a shared
incrementing counter between all the macro instantiations. Oh and also
it limits us to 65537 registrations.

Furthermore this requires users to enroll in feature(asm_experimental_arch).

As such this approach feels doubly risky since wasm is not a supported
platform for `global_asm!` and we are clearly circumventing rusts
unwillingness to let us annotate statics with a custom section in wasm
by simply emitting the directives to the assembler directly.

(but perhaps these are both just conservative restrictions anyway?)